Movement

+++ Assets/Scripts/Pathfinding/Path.cs
@@ -2,10 +2,10 @@ class Path {
 	public int g; //Steps from A to this
 	public int h; //Steps from this to B
 	public Path parent;
-	public int x;
-	public int y;
+	public float x;
+	public float y;
 
-	public Path(int _g, int _h, Path _parent, int _x, int _y) {
+	public Path(int _g, int _h, Path _parent, float _x, float _y) {
 		g = _g;
 		h = _h;
 		parent = _parent;
